refactor: backtrack to use usize instead of i32 for start and n

+++ 140-word-break-II/main.rs
@@ -6,20 +6,20 @@ impl Solution {
         }
 
         let mut sentences = Vec::new();
-        let n = s.len() as i32;
+        let n = s.len();
 
         Self::backtrack(0, &mut Vec::new(), &s, n, &word_set, &mut sentences);
         sentences
     }
 
-    fn backtrack(start: i32, path: &mut Vec<String>, s: &String, n: i32, word_set: &std::collections::HashSet<String>, sentences: &mut Vec<String>) {
+    fn backtrack(start: usize, path: &mut Vec<String>, s: &String, n: usize, word_set: &std::collections::HashSet<String>, sentences: &mut Vec<String>) {
         if start == n {
             sentences.push(path.join(" "));
             return;
         }
 
         for i in start..n {
-            let word = &s[start as usize..i as usize + 1];
+            let word = &s[start..i+ 1];
             if word_set.contains(word) {
                 path.push(word.to_string());
                 Self::backtrack(i + 1, path, s, n, word_set, sentences);
